server_python/requirements.txt
fastapi
pydantic
mysql-connector-python
uvicorn